+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Match `Cache-Control: no-transform` directive case-insensitively
+
 1.8.1 / 2025-07-17
 ==========